[ParkingLot][Bryan] refactor: made getAvailableCapacity as int and cast it to double in getAvailablePositionRate

+++ src/main/java/com/oocl/cultivation/ParkingLot.java
@@ -45,11 +45,11 @@ public class ParkingLot {
         return capacity;
     }
 
-    double getAvailableCapacity() {
+    int getAvailableCapacity() {
         return getCapacity() - getParkedCarsSize();
     }
 
     double getAvailablePositionRate() {
-        return getAvailableCapacity() / capacity;
+        return (double) getAvailableCapacity() / capacity;
     }
 }
